[level 1] Title: [1차] 비밀지도, Time: 7.30 ms, Memory: 79.2 MB -BaekjoonHub

+++ 프로그래머스/lv1/17681. ［1차］ 비밀지도/［1차］ 비밀지도.java
@@ -3,14 +3,15 @@ class Solution {
         String[] answer = new String[n];
 
         for (int i = 0; i < n; i++) {
-
-            String binaryString = Long.toBinaryString(arr1[i] | arr2[i]); 
-
+            
+            String binaryString = Integer.toBinaryString(arr1[i] | arr2[i]);
             binaryString = String.format("%" + n + "s", binaryString);
-
-            binaryString = binaryString.replace("1", "#");
-            binaryString = binaryString.replace("0", " ");
+            
+            binaryString = binaryString.replace("1" , "#");
+            binaryString = binaryString.replace("0" , " ");
+            
             answer[i] = binaryString;
+
         }
 
         return answer;
